feat: Header 컴포넌트 분리, loading 추가, Avatar showFallback 설정
- 공통 컴포넌트 안 컴포넌트 분리를 위해 폴더 생성
