+++ src/app/(auth)/my-page/profile/page.tsx
@@ -172,6 +172,7 @@ const MyProfile = () => {
             <div className="mx-auto flex h-full max-w-md flex-col gap-4 p-4">
               <div className="flex items-center">
                 <Avatar
+                  showFallback
                   className="mr-4 h-20 w-20 text-large"
                   alt="프로필 사진"
                   src={previewUrl || user.imageUrl}
