+++ src/app/components/header/Header.tsx
@@ -1,0 +1,36 @@
+'use client';
+
+import React from 'react';
+import { useQuery } from '@tanstack/react-query';
+import { postTokenRefresh } from '@/services/token/postTokenRefresh';
+import Logo from './Logo';
+import GoMyPage from './GoMyPage';
+import GoLogin from './GoLogin';
+
+const Header = () => {
+  const { data: token } = useQuery({
+    queryKey: ['tokenData'],
+    queryFn: postTokenRefresh,
+  });
+
+  if (!token) {
+    return (
+      <div className="sticky top-0 z-30 flex h-[61px] w-full max-w-[600px] items-center justify-between border-b-1 bg-background pl-4">
+        <Logo />
+        <div className="flex items-center gap-2 pr-4">
+          <GoLogin />
+        </div>
+      </div>
+    );
+  }
+
+  return (
+    <div className="sticky top-0 z-30 flex h-[61px] w-full max-w-[600px] items-center justify-between border-b-1 bg-background pl-4">
+      <Logo />
+      <div className="flex items-center gap-2 pr-4">
+        <GoMyPage />
+      </div>
+    </div>
+  );
+};
+export default Header;
